refactor: move session socket adapters to cloudflare
