+++ packages/control-plane/src/session/durable-object.ts
@@ -68,8 +68,8 @@ import { WsClientMappingRepository } from "./ws-client-mapping-repository";
 import { resolveParticipantName } from "./participant-name";
 import { validateReasoningEffort } from "./reasoning-effort";
 import { parseTunnelUrls } from "./tunnel-urls";
-import { DurableObjectSocketRegistry } from "./durable-object-socket-registry";
-import { DurableObjectSessionConnections } from "./durable-object-session-connections";
+import { DurableObjectSocketRegistry } from "../cloudflare/durable-object-socket-registry";
+import { DurableObjectSessionConnections } from "../cloudflare/durable-object-session-connections";
 import { SessionPullRequestStore } from "../db/session-pull-request-store";
 import { PullRequestCreationClaims, SessionPullRequestService } from "./pull-request-service";
 import { refreshSessionPullRequests } from "./pull-request-refresh";
